requirements.txt
diffusers==0.36.0
facenet_pytorch==2.6.0
lpips==0.1.4
matplotlib==3.10.8
numpy==2.4.4
opencv_python==4.13.0.90
pandas==3.0.1
Pillow==12.1.1
PyYAML==6.0.3
scipy==1.17.1
seaborn==0.13.2
skimage==0.0
streamlit==1.55.0
tensorflow==2.21.0
torch==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
